requirements.txt
appdirs==1.4.4
Django==2.2.15
djangorestframework==3.11.1
Jinja2==2.11.2
virtualenv==20.0.30
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==1.2.0
